% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{odds_to_prob}
\alias{odds_to_prob}
\title{convert odds to probability}
\usage{
odds_to_prob(odds)
}
\arguments{
\item{odds}{A vector of odds}
}
\value{
The probablity
}
\description{
\code{odds_to_prob} covert odds to probabilty
}
\examples{
odds_to_prob(c(2.0,5.0,11.2))
}

